Make move generator array in fruiton modifiable
I feel this is how we want to do things. If some effect adds movement capability to fruiton given fruiton
has it until something removes it. Even if we later decide we want to do it differently this gives us more
options since the design is less restrictive than using ReadOnlyArray. (We can use it later if we want to).